+++ src/android/com/outsystems/plugins/healthfitness/OSHealthFitness.kt
@@ -13,6 +13,7 @@ import com.outsystems.osnotificationpermissions.*
 import com.outsystems.plugins.healthfitness.background.BackgroundJobParameters
 import com.outsystems.plugins.healthfitness.background.DatabaseManager
 import com.outsystems.plugins.healthfitness.background.UpdateBackgroundJobParameters
+import com.outsystems.plugins.healthfitness.data.types.HealthAdvancedQueryParameters
 import com.outsystems.plugins.healthfitness.store.*
 import com.outsystems.plugins.oscordova.CordovaImplementation
 import org.apache.cordova.*
@@ -170,8 +171,8 @@ class OSHealthFitness : CordovaImplementation() {
     }
 
     private fun advancedQuery(args: JSONArray) {
-        val parameters = gson.fromJson(args.getString(0), AdvancedQueryParameters::class.java)
-        healthStore?.advancedQueryAsync(
+        val parameters = gson.fromJson(args.getString(0), HealthAdvancedQueryParameters::class.java)
+        healthConnectViewModel.advancedQuery(
             parameters,
             { response ->
                 val pluginResponseJson = gson.toJson(response)
@@ -186,11 +187,11 @@ class OSHealthFitness : CordovaImplementation() {
     private fun writeData(args: JSONArray) {
         try {
             val variable = args.getString(0)
-            val healthRecordName = HealthRecordName.valueOf(variable)
+            val healthRecord = HealthRecord.valueOf(variable)
             val value = args.getDouble(1)
 
             healthConnectViewModel.writeData(
-                healthRecordName,
+                healthRecord,
                 value,
                 getActivity().packageName,
                 {
@@ -208,7 +209,7 @@ class OSHealthFitness : CordovaImplementation() {
     private fun getLastRecord(args: JSONArray) {
         try {
             healthConnectViewModel.getLastRecord(
-                HealthRecordName.valueOf(args.getString(0)),
+                HealthRecord.valueOf(args.getString(0)),
                 {
                     sendPluginResult(it, null)
                 },
@@ -252,14 +253,14 @@ class OSHealthFitness : CordovaImplementation() {
     }
 
     private fun deleteBackgroundJob(args: JSONArray) {
-        val parameters = args.getString(0)
-        healthStore?.deleteBackgroundJob(
-            parameters,
-            { response ->
-                sendPluginResult(response)
+        val jobId = args.getString(0)
+        healthConnectViewModel.deleteBackgroundJob(
+            jobId,
+            {
+                sendPluginResult("success", null)
             },
-            { error ->
-                sendPluginResult(null, Pair(error.code.toString(), error.message))
+            {
+                sendPluginResult(null, Pair(it.code.toString(), it.message))
             }
         )
     }
@@ -278,13 +279,13 @@ class OSHealthFitness : CordovaImplementation() {
 
     private fun updateBackgroundJob(args: JSONArray) {
         val parameters = gson.fromJson(args.getString(0), UpdateBackgroundJobParameters::class.java)
-        healthStore?.updateBackgroundJob(
+        healthConnectViewModel.updateBackgroundJob(
             parameters,
-            { response ->
-                sendPluginResult(response)
+            {
+                sendPluginResult("success", null)
             },
-            { error ->
-                sendPluginResult(null, Pair(error.code.toString(), error.message))
+            {
+                sendPluginResult(null, Pair(it.code.toString(), it.message))
             }
         )
     }